refactor: move logic to domain, prepare tests for tributes
